feat: resolve function in using for
