+++ src/main/kotlin/me/serce/solidity/lang/resolve/ref/refs.kt
@@ -20,6 +20,9 @@ class SolUserDefinedTypeNameReference(element: SolUserDefinedTypeName) : SolRefe
     val parent = element.parent
     if (parent is SolNewExpressionElement) {
       return SolResolver.resolveNewExpression(parent)
+    } else if (parent is SolUsingForElement) {
+      val resolved = SolResolver.resolveUsingForElement(element)
+      return if (resolved != null) listOf(resolved) else emptyList()
     }
     return SolResolver.resolveTypeNameUsingImports(element)
   }
