+++ src/test/kotlin/me/serce/solidity/lang/core/resolve/SolFunctionResolveTest.kt
@@ -570,22 +570,6 @@ class SolFunctionResolveTest : SolResolveTestBase() {
     }
   """)
 
-  fun testResolveFunctionInUsingFor() = checkByCode("""
-    pragma solidity ^0.8.26;
-    
-    type Foo is uint256;
-
-    function isHappy(Foo f) internal pure returns(bool) {
-                //x   
-        return Foo.unwrap(f) > 100;
-    }
-    
-    using {
-        isHappy
-          //^
-    } for Foo global;
-     """)
-
   fun testResolveFunctionWithUsingForInImportedFile() = testResolveBetweenFiles(
     InlineFile(
       code = """
