+++ src/test/kotlin/me/serce/solidity/lang/core/resolve/SolUserDefinedValueTypeResolveTest.kt
@@ -69,4 +69,38 @@ class SolUserDefinedValueTypeResolveTest : SolResolveTestBase() {
     """
     )
   )
+
+  fun testResolveFunctionInUsingFor() = checkByCode("""
+    pragma solidity ^0.8.26;
+    
+    type Foo is uint256;
+
+    function isHappy(Foo f) internal pure returns(bool) {
+                //x   
+        return Foo.unwrap(f) > 100;
+    }
+    
+    using {
+        isHappy
+          //^
+    } for Foo global;
+     """)
+
+  fun testResolveFunctionInUsingFor2() = checkByCode("""
+    pragma solidity ^0.8.26;
+    
+    type Foo is uint256;
+
+    library FooLib {
+        function isHappy(Foo f) internal pure returns(bool) {
+                    //x   
+            return Foo.unwrap(f) > 100;
+        }
+    }
+    
+    using {
+        FooLib.isHappy
+                //^
+    } for Foo global;
+     """)
 }
